# swagger/ecommerceapi.yaml
openapi: 3.0.3
info:
  title: Ecommerce Portfolio Project (part one)
  version: 1.0.0
  description: Backend (part one) of the Codecademy ecommerce portfolio project
paths:
...
  /product/{id}:
    summary: Get product by id.
    description: Retrieves the product with a specific id.
    get:
      summary: Get product by id.
      description: Retrieves product with the given id.
      operationId: get_product_by_id
      parameters:
        - name: id
          in: path
          description: ID of the product
          required: true
          schema:
            type: integer
      responses:
        default:
          description: Default error sample response
      tags:
        - Products
  /product/categories/{category}:
    summary: Get products by category.
    description: Retrieves products by a given category.
    get:
      summary: Get products by category.
      description: Retrieves products by a given category.
      operationId: get_products_by_category
      parameters:
        - name: category
          in: path
          description: Category of the product
          required: true
          schema:
            type: string
      responses:
        default:
          description: Default error sample response
      tags:
        - Products
  /product/delete/{id}:
    summary: Delete product by id,
    description: Retrieves product by a given id
    delete:
      summary: Delete product by id.
      description: Deletes poduct by a given id.
      operationId: delete_product_by_id
      parameters:
        - name: id
          in: path
          description: Id of the product
          required: true
          schema:
            type: string
      responses:
        default:
          description: Default error sample response
      tags:
        - Products
  /product/addProduct:
    summary: Add a new product
    description: Add a new product.
    post:
      summary: Add a new product
      description: Add a new product
      operationId: add_new_product
      responses:
        default:
          description: Default error sample response
      tags:
        - Products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                price:
                  type: number
                description:
                  type: string
                category:
                  type: string
            examples:
              example1:
                value:
                  name: "Socks 6"
                  price: 7.99
                  description: "Lightweight and breathable socks"
                  category: "socks"
  /product/updateProduct/{id}:
    summary: Update product by id
    description: Update product by given id.
    put:
      summary: Update existing product
      description: Update existing product by given id
      operationId: update_product
      responses:
        default:
          description: Default error sample response
      parameters:
        - name: id
          in: path
          description: Id of the product
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                price:
                  type: number
                description:
                  type: string
                category:
                  type: string
            examples:
              example1:
                value:
                  name: "Updated Socks"
                  price: 9.99
                  description: "Updated description"
                  category: "updated-category"
      tags:
        - Products
...